Moved config file loading to its own file
This also includes a structure class that acts as a container for the
IRC parameters passed to the IRC object. Daemon plugin was changed to
reflect this.

Structures file will act as a global structure class holder. IRC's
Message class was moved to this file. Anything that needs to create
messages needs to import structures or IRC. Plugins require the IRC
import so they should have access to it automatically.