Added Help plugin and minor bug fix
Added the help plugin that lists public commands of all plugins.

Fixed a minor bug where if a user does not specify a debug level, debug
is not is not defined. Now minimal debug info is the default.

IRC class now contains an internal list of initiated plugins, set by the
main file after IRC initialization.