Major update: Users are now handled with SQLite
Replaced users_db.py with users_sqlite. All user tracking is done
through SQLite now. This module is still pretty fragile and likely still
has a major bug or two in it.

Converted all plugins to the new user system.

Creep plugin added, will give information on registered users.

Fixed bug in the parser when a user joins a channel where it wasn't
stripping the colon. This caused the wrong channel to be created.

Fixed bug in Help plugin, causing the plugin list to double everytime it
was called. Changed plugin finding code to run only once.

Added message listener to the user_manager plugin, will now try to
register nicks to users when they talk in a channel.

Added nick change to the parser and is now a valid message type.

Added mode change to the parser (just changing user statuses in a
channel) and message type. User status changes are now properly
recorded.

Kicks are now part of the parser/message type.
- Auto re-join is now implemented for the bot