Added ChatLog plugin
Fixed crash in change_user_status in the SQLite file. This whole fuke
needs to be refactored.

Also added am_i_op method the the SQLite file, plugins should be able to
determine if the bot's opped before continuing their work. Unfortunately
this often fires before the user status is changed in the database. It's
finally come to the point where it's going to be necessary to implement
internal plugins and external, with internal getting messages first.

Added the ChatLog plugin, will record the user id, channel id, a
timestamp and the chat message on all "message" type of messages (this
currently includes /me actions, don't know how I feel about that yet).